requirements.txt
requests
flask-lambda-python36
boto3
pandas
datetime
calendar
croniter